backend/requirements.txt
# FastAPI & server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
slowapi>=0.1.9

# HTTP & resilience
httpx>=0.26.0
tenacity>=8.2.0

# Config & validation
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Database (PostgreSQL or SQLite for local)
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
aiosqlite>=0.19.0
greenlet>=3.0.0

# Parsing & extraction
beautifulsoup4>=4.12.0
trafilatura>=1.6.0
readability-lxml>=0.8.1
lxml>=5.0.0

# LLM (OpenAI-compatible)
openai>=1.12.0

# Optional browser fallback
playwright>=1.41.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
